requirements.txt
pandas
ShopifyAPI
requests